Move _Py_AtExit() to the public API.
